requirements.txt
streamlit
torch
torchvision
diffusers
transformers
Pillow
numpy
